v4.9 - clean up habit controller
