+++ controllers/habits.js
@@ -1,6 +1,15 @@
 const Habit = require('../models/habit');
-const User = require('../models/user');
 const HabitLog = require('../models/habitLog');
+
+const { 
+	autoTrackHabit,
+	enableDisableTracking,
+	updateLastAutotracked,
+	updateStreak,
+	updateStatus,
+	getHistory,
+	getNextDate
+} = require('../utils/habitUtils');
 
 const frequencies = ['daily']; 
 
@@ -59,129 +68,6 @@ const showHabit = async (req, res) => {
 	}
 };
 
-async function autoTrackHabit (habit) {
-	// check if habit due date has passed
-	let currentDate = new Date();
-	let cutoffDate = new Date(currentDate);
-	
-	cutoffDate.setDate(currentDate.getDate() - 2);
-	// let dueDate = habit.due;
-	console.log("Cutoff:", cutoffDate);
-
-	// While the due date is before cutoff date
-	while (cutoffDate > habit.due) {
-		console.log('updating missed due dates');
-		// 1. update habit log entry for due date -> update status to missed
-		let updatedLog = await HabitLog.findByIdAndUpdate( 
-			habit.last_log,
-			{ status: 'missed' }, 
-			{ new: true, useFindAndModify: false, runValidators: true }
-		);
-		
-		console.log("Updated log:", updatedLog);
-
-		// 2. update due to next due and interval next_due
-		habit.due = habit.next_due;
-		habit.next_due = getNextDate(habit.due, 1);
-
-		// 3. Add a new entry in habit log with the new due date
-		let newLog = new HabitLog({
-			habit: habit.id,
-			date: habit.due,
-			status: 'pending'
-		});
-
-		habit.last_log = newLog.id;
-		console.log("new log:", newLog);
-		console.log("Updated habit:", habit);
-
-		// 4. Save changes to DB
-		await habit.save();
-		await newLog.save();
-	}
-
-	await updateStreak(habit);
-	await updateStatus(habit);
-	return habit;
-}
-
-async function enableDisableTracking(habit) {
-	let currentDate = new Date();
-	if (habit.due > currentDate && habit.due.getDate()!== currentDate.getDate()) {
-		habit.is_tracking_enabled = false;
-	} else {
-		habit.is_tracking_enabled = true;
-	}
-	await habit.save();
-	return habit;
-}
-
-async function updateLastAutotracked(userId, currentDate) {
-	await User.findByIdAndUpdate(
-		userId,
-		{ last_autotracked: currentDate },
-		{ new: true, useFindAndModify: false, runValidators: true }
-	);
-}
-
-async function updateStreak (habit) {
-	// Get habit logs with status 'missed/complete' ordered in reverse order of date
-	const habitLogs = await HabitLog.find({ 
-		habit: habit.id, 
-		status: { $in: ["missed", "complete"] } 
-	}).sort({ date: 'desc' });
-	
-	let streak = 0;
-	for (let log of habitLogs) {
-		if (log.status === 'complete') {
-			streak++;
-		} else {
-			break;
-		}
-	}
-
-	habit.streak = streak;
-	await habit.save();
-	return habit;
-}
-
-async function getHistory (habit) {
-	const history = await HabitLog.find({
-		habit: habit.id, 
-		status: { $in: ["missed", "complete"] } 
-	})
-	.limit(30)
-	.sort({ date: 'desc' });
-
-	return history;
-}
-
-async function updateStatus (habit) {
-	const last7days = await HabitLog.find({
-		habit: habit.id, 
-		status: { $in: ["missed", "complete"] } 
-	})
-	.limit(7)
-	.sort({ date: 'desc' });
-
-	let missedCount = 0;
-
-	for (let log of last7days) {
-		if (log.status === "missed") missedCount++;
-	};
-
-	if (missedCount === 0 ) {
-		habit.status = "On track";
-	} else if (missedCount <= 2 ) {
-		habit.status = "Could be better";
-	} else {
-		habit.status = "Off the rails";
-	}
-
-	await habit.save();
-	return habit;
-}
-
 const renderNewHabitForm = (req, res) => {
 	res.render('habits/new', { frequencies });
 };
@@ -209,13 +95,6 @@ const createHabit = async (req, res) => {
 		res.redirect('/habits/new');
 	}
 };
-
-// move to utils
-function getNextDate(date, interval) {
-	let nextDate = new Date(date);
-	nextDate.setDate(date.getDate() + interval);
-	return nextDate;
-}
 
 const renderEditHabitForm = async (req, res) => {
 	try {
@@ -295,7 +174,6 @@ const trackHabit = async (req, res) => {
 		res.redirect(`/habits/${habit.id}`);
 	}
 };
-
 
 module.exports = {
 	renderDashboard,
